week04-homework/requirements.txt
langchain-core
langchain-deepseek
langgraph
pydantic
pydantic-settings
redis
loguru
python-dotenv

